Putting API-key-related things into their own file and folder
